+++ src/layer_tools/gradient_computer.py
