+++ src/utils/network_gradient_comparator.py
@@ -8,10 +8,10 @@ from typing import List, Tuple
 
 import numpy as np
 
-from src.neural_network.error_functions import ErrorFunctionLike
-from src.neural_network.layer_interfaces import WeightsHavingLayerLike
+from src.network_functions.error_functions import ErrorFunctionLike
+from src.layer_tools.layer_interfaces import WeightsHavingLayerLike
 from src.neural_network.neural_network import NetworkLayerManager
-from src.utils.data_processor import Dataset
+from src.data_processing.dataset import Dataset
 
 
 class NetworkGradientComparator:
